tests/test_api.py
# tests/test_api.py
import pytest


def test_placeholder():
    """Placeholder test - add real tests as needed."""
    assert True


def test_model_names():
    """Test that model names are valid."""
    valid_models = ["logistic", "lightgbm", "xgboost"]
    for model in valid_models:
        assert isinstance(model, str)
        assert len(model) > 0
