requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
python-dotenv==1.0.0

# ML
joblib==1.3.2
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
lightgbm==4.2.0
xgboost==2.0.3

# Dashboard
streamlit==1.31.0
requests==2.31.0

# Streaming
kafka-python==2.0.2

# Monitoring
prometheus-client==0.19.0

# Database
sqlalchemy==2.0.25
